add detect to separate directory
change dir structure of detect api and add empty test case file for detect api
